refactor source classes structure and create Main.java
