+++ src/main/java/com/gridnine/testing/Main.java
@@ -1,0 +1,8 @@
+package com.gridnine.testing;
+
+public class Main {
+
+    public static void main(String[] args) {
+    }
+
+}
