+++ src/main/java/com/gridnine/testing/flight/Flight.java
@@ -1,0 +1,25 @@
+package com.gridnine.testing.flight;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+/**
+ * Bean that represents a flight.
+ */
+public class Flight {
+    private final List<Segment> segments;
+
+    Flight(final List<Segment> segs) {
+        segments = segs;
+    }
+
+    public List<Segment> getSegments() {
+        return segments;
+    }
+
+    @Override
+    public String toString() {
+        return segments.stream().map(Object::toString)
+                .collect(Collectors.joining(" "));
+    }
+}
